requirements.txt
Django==3.2.3
django-crispy-forms == 1.12.0
django-allauth == 0.45.0
django-debug-toolbar == 3.2.2
django-model-utils == 4.1.1
django-autoslug == 1.9.8
Pillow == 8.3.1
requests==2.25.1
